% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{cdf,ContRV-method}
\alias{cdf,ContRV-method}
\title{Generarea CDF-ului unei variabile aleatoare continue}
\usage{
\S4method{cdf}{ContRV}(crv)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă}
}
\value{
Returnează o funcție care reprezintă cdf-ul obiectului trimis ca parametru.
}
\description{
Metodă a clasei \code{ContRV} care genereză funcția de repartiție (funcția
de distribuție cumulativă, cdf) din funcția densitate de probabilitate a
variabilei aleatoare continue.
}
\details{
Pentru utilizare corectă în unele cazuri trebuie utilizată funcția Vectorize asupra funcției cdf obținute
}
\examples{
var1 <- ContRV(pdf = function(x) { 1 / 3 }, lowerBound = 3, upperBound = 6)
var1_cdf <- cdf(var1)

crv <- ContRV(pdf = function(x) { 2 * exp(-2 * x) }, lowerBound = 0, upperBound = Inf)
crv_cdf <- cdf(crv)
}
