% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{cdf}
\alias{cdf}
\title{Funcția generică cdf}
\usage{
cdf(crv)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}
}
\value{
Returnează o funcție care reprezintă cdf-ul obiectului trimis ca parametru.
}
\description{
Funcție generică disponibilă pentru overloading
}
