% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{check_if_pdf}
\alias{check_if_pdf}
\title{Verificarea unei funcții dacă este funcție densitate de probabilitate}
\usage{
check_if_pdf(func, lowerBound = -Inf, upperBound = Inf)
}
\arguments{
\item{func}{Funcția care trebuie să o verificăm dacă este pdf}

\item{lowerBound}{Limita inferioară a intervalului specificat (default: \code{-Inf})}

\item{upperBound}{Limita superioară a intervalului specificat (default: \code{Inf})}
}
\value{
Returnează \code{TRUE} dacă funcția introdusă este pdf, \code{FALSE} dacă nu este
și eroare dacă integrala este divergentă
}
\description{
Fiind dată o funcție și limita superioara și inferioară,
verifică dacă aceasta este o funcție densitate de probabilitate (pdf), verificând dacă
respectă proprietățile acesteia:
\itemize{
    \item Funcția este nenegativă pe intervalul \code{[lowerBound, upperBound]}
    \item Integrala funcției pe acest interval este egală cu 1.
}
}
\details{
Erori la verificarea funcției pot apărea în cazul în care integrala este divergentă pe acel interval
}
\examples{
fun <- function(x) { 1 }
check_if_pdf(fun, 0, 1)   # FALSE

fun <- function(x) { 2 * exp(-2*x) }
check_if_pdf(fun, 0, Inf)   # TRUE
}
