% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{chi}
\alias{chi}
\title{Repartiția Chi}
\usage{
chi(k)
}
\arguments{
\item{k}{Gradul de libertate}
}
\value{
Returnează un obiect de tip \code{ContRV}.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție Chi
}
\details{
Repartiția Chi are următoarele caracteristici:
\itemize{
    \item Parametri: \code{k > 0}
    \item Domeniul de valori: \code{[0, Inf)}
    \item Densitate: \eqn{ f(x) = (x^(k - 1) * exp( -(x^2) / 2)) / (2^((k / 2) - 1) * gamma(k / 2))}
    \item Aplicatie: Repartiția Chi poate reprezenta distribuția distanțelor Euleriene a unui set de puncte distribuite dupa repartiția normala.
}
Funcția returnează o eroare în cazul în care \eqn{k <= 0}
}
\examples{
var1 <- chi(3)
# var2 <- chi(-1) # Eroare: -1 <= 0
}
\references{
Chi distribution: https://en.wikipedia.org/wiki/Chi_distribution
}
