% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{conditional_density_Y_X,ContRV,ContRV,function-method}
\alias{conditional_density_Y_X,ContRV,ContRV,function-method}
\title{Densitatea lui Y condiționată de X din doua variabile aleatoare continue}
\usage{
\S4method{conditional_density_Y_X}{ContRV,ContRV,`function`}(crv1, crv2, dens)
}
\arguments{
\item{crv1}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{crv2}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{dens}{Funcția densitate comună a celor două variabile aleatoare}
}
\value{
Returnează o funcție, ce reprezintă densitatea lui Y condiționată de X.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează densitatea lui Y condiționată de X,
din două variabile aleatoare continue, utilizând funcția densitate comună acelor două.
}
\examples{
obj1 <- ContRV(pdf = function(x) { 1 }, lowerBound = 0, upperBound = 1)
obj2 <- ContRV(pdf = function(x) { 1 }, lowerBound = 1, upperBound = 2)
f3 <- conditional_density_Y_X(obj1, obj2, function(x, y) 8/3 * x^3 *y)
f3(3, 4) # 2.6667
}
