% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{conditional_density_Y_X}
\alias{conditional_density_Y_X}
\title{Funcția generică conditional_density_Y_X}
\usage{
conditional_density_Y_X(crv1, crv2, dens)
}
\arguments{
\item{crv1}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{crv2}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{dens}{Funcția densitate comună a celor două variabile aleatoare}
}
\value{
Returnează o valoare care reprezintă varianța obiectului trimis ca parametru.
}
\description{
Funcție generică disponibilă pentru overloading
}
