% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\docType{class}
\name{ContRV-class}
\alias{ContRV-class}
\alias{ContRV}
\title{Variabilă aleatoare continuă}
\value{
Returnează un obiect de tip variabilă aleatoare continuă sau eroare.
}
\description{
Clasă pentru variabile aleatoare continue. Pentru generarea unui obiect de tip
\code{ContRV}, e nevoie de o funcție densitate de probabilitate și limitele superioare și inferioare
ale funcției
}
\details{
Erori la crearea unui obiect pot apărea în cazul în care funcția introdusă nu este funcție densitate
de probabilitate
}
\section{Slots}{

\describe{
\item{\code{pdf}}{Funcție densitate de probabilitate}

\item{\code{lowerBound}}{Limita inferioară a intervalului specificat}

\item{\code{upperBound}}{Limita superioară a intervalului specificat}
}}

\examples{
var1 <- ContRV(pdf = function(x) { 1 / 3 }, lowerBound = 3, upperBound = 6)
var2 <- ContRV(pdf = function(x) { 2 * exp(-2 * x) }, lowerBound = 0, upperBound = Inf)
}
