% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{-,ContRV,ContRV-method}
\alias{-,ContRV,ContRV-method}
\title{Scăderea a două variabile aleatoare continue independente}
\usage{
\S4method{-}{ContRV,ContRV}(e1, e2)
}
\arguments{
\item{e1}{Obiect de tip variabilă aleatoare continuă}

\item{e2}{Obiect de tip variabilă aleatoare continuă}
}
\value{
Returnează un obiect de tip variabilă aleatoare continuă care reprezintă diferența
dintre prima și a doua varibilă.
}
\description{
Operație de scădere pentru a două obiecte de tip \code{ContRV}. Scăderea se face
folosind formula de convoluție, ambele variabile primite ca parametru fiind variabilie
aleatoare continue independente.
}
\details{
Folosește formula de convoluție pentru scăderea celor două variabile.
}
\examples{
fun1 <- normalize_function(function(x) { x^2 }, -2, 10)
var1 <- ContRV(pdf = fun1, lowerBound = -2, upperBound = 10)

fun2 <- normalize_function(function(x) {x + 1}, 0, 10)
var2 <- ContRV(pdf = fun2, lowerBound = 0, upperBound = 10)

var_sum <- var1 - var2
}
