% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{Cor,ContRV,ContRV,function-method}
\alias{Cor,ContRV,ContRV,function-method}
\title{Coeficientul de corelație a doua variabile aleatoare continue}
\usage{
\S4method{Cor}{ContRV,ContRV,`function`}(crv1, crv2, dens)
}
\arguments{
\item{crv1}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{crv2}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{dens}{Funcția densitate comună a celor două variabile aleatoare}
}
\value{
Returnează o valoare numerică, ce reprezintă coeficientul de corelație ale celor două
variabilelor aleatoare continue.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează coeficientul de corelație a două variabile
aleatoare continue, utilizând funcția densitate comună acelor două.
}
\details{
Utilizează covarinața celor două variabile pentru calcularea coeficientului de corelanță.
}
\examples{
obj1 <- ContRV(pdf = function(x) { x }, lowerBound = 0, upperBound = sqrt(2))
obj2 <- ContRV(pdf = function(x) { 1 }, lowerBound = 0, upperBound = 1)
Cor(obj1, obj2, function(x, y) 4*x*y)
}
