% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{Cov,ContRV,ContRV,function-method}
\alias{Cov,ContRV,ContRV,function-method}
\title{Covarianța a doua variabile aleatoare continue}
\usage{
\S4method{Cov}{ContRV,ContRV,`function`}(crv1, crv2, dens)
}
\arguments{
\item{crv1}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{crv2}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{dens}{Funcția densitate comună a celor două variabile aleatoare}
}
\value{
Returnează o valoare numerică, ce reprezintă covarianța variabilelor aleatoare continue.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează covarianța a două variabile
aleatoare continue, utilizând funcția densitate comună acelor două.
}
\details{
Utilizează integrala dublă pentru calcularea covarianței celor două variabile aleatoare.
}
\examples{
obj1 <- ContRV(pdf = function(x) { x }, lowerBound = 0, upperBound = sqrt(2))
obj2 <- ContRV(pdf = function(x) { 1 }, lowerBound = 0, upperBound = 1)
Cov(obj1, obj2, function(x, y) 4*x*y)
}
