% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{EPSILON}
\alias{EPSILON}
\title{Eroarea aboslută}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
EPSILON
}
\description{
Setează o valoare constantă pentru eroarea abosolută
}
\details{
Această constantă este specifică acestui proiect
}
\keyword{datasets}
