% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{exponential}
\alias{exponential}
\title{Repartitia exponențială}
\usage{
exponential(lambda)
}
\arguments{
\item{lambda}{Parametrul repartiției exponențiale (rata)}
}
\value{
Returnează un obiect de tip \code{ContRV} sau o eroare.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție exponențială.
}
\details{
Repartiția exponențială are următoarele caracteristici:
\itemize{
    \item Parametru: \code{lambda > 0}
    \item Domeniul de valori: \code{[0, Inf]}
    \item Densitate: f(x) = lambda * exp(-lambda * x)
    \item Repartiția: F(x) = 1 - exp^(-(lambda * x))
    \item Mean: 1 / lambda
    \item Mediana: ln(2) / lambda
    \item Dispersia: 1 / lambda^2
}
Funcția returnează o eroare în cazul în care \code{lambda <= 0} \cr
Repartiția exponențială este deseori întâlnită în domenii precum: fizica
(timpul de descomunere radioactivă a particulelor), biologia (distanța dintre
mutații de ADN). De asemenea, se pot întâlni și în viața de zi cu zi: timpul până
la următorul apel telefonic.
}
\examples{
var1 <- exponential(2)
# var2 <- exponential(-1) # Eroare: -1 <= 0
}
\references{
Exponential distribution: https://en.wikipedia.org/wiki/Exponential_distribution
}
