% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{generate_crv,ContRV,numeric-method}
\alias{generate_crv,ContRV,numeric-method}
\title{Generarea valorilor dintr-o repartiție de variabile aleatoare continue}
\usage{
\S4method{generate_crv}{ContRV,numeric}(crv, n)
}
\arguments{
\item{crv}{Obiectul de tip variabilă aleatoare continuă}

\item{n}{Numarul de valori pe care dorim să le generăm}
}
\value{
Returnează cele n valori generate.
}
\description{
Funcție care generează n valori dintr-o repartiție de variabile aleatoare continue, utilizând simularea.
}
\details{
Utilizează simularea pentru generarea valorilor.
}
\examples{
generate_crv(exponential(4), 10)
generate_crv(uniform(1, 5), 15)
generate_crv(normal(0, 1), 20)
}
