% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{generate_crv}
\alias{generate_crv}
\title{Funcția generică generate_crv}
\usage{
generate_crv(crv, n)
}
\arguments{
\item{crv}{Funcția de repartiție a unei variabile aleatoare continue}

\item{n}{Numarul de valori pe care dorim să le generăm}
}
\description{
Funcție generică disponibilă pentru overloading
}
