% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{logNormal}
\alias{logNormal}
\title{Repartiția Lognormală}
\usage{
logNormal(mean, disp)
}
\arguments{
\item{mean}{Media variabilei aleatoare}

\item{disp}{Dispersia varibilei aleatoare}
}
\value{
Returnează un obiect de tip \code{ContRV} sau o eroare.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție log-normală.
}
\details{
Repartiția uniformă are următoarele caracteristici:
\itemize{
    \item Parametri: \code{mean, disp > 0}
    \item Domeniul de valori: \code{(o, Inf)}
    \item Densitate: \eqn{f(x) = 1 / (sd * x * sqrt(2 * pi)) * exp(-((log(x) - mean)^2)/( 2 * disp))}
    \item Aplicație: Dacă Variabila aleatoare \eqn{X} are repartiție log-normală, atunci \eqn{Y = ln(X)} este o repartiție normală. Repartiția log-normală se utilizează în simulări care se comportă ca și repartițiile normale, însă nu pot lua valori negative.
}
Funcția returnează o eroare în cazul în care \eqn{disp <= 0}
}
\examples{
var1 <- logNormal(0, 1)
# var2 <- logNormal(0, -100) # Eroare: -100 <= 0
}
\references{
Lognormal distribution: https://en.wikipedia.org/wiki/Log-normal_distribution
}
