% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{marginal_pdf_1}
\alias{marginal_pdf_1}
\title{Funcția generică marginal_pdf_1}
\usage{
marginal_pdf_1(crv1, crv2, dens)
}
\arguments{
\item{crv1}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{crv2}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{dens}{Funcția densitate comună a celor două variabile aleatoare}
}
\value{
Returnează o valoare care reprezintă varianța obiectului trimis ca parametru.
}
\description{
Funcție generică disponibilă pentru overloading
}
