% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{marginal_pdf_2,ContRV,ContRV,function-method}
\alias{marginal_pdf_2,ContRV,ContRV,function-method}
\title{Densitatea marginală al lui y din doua variabile aleatoare continue}
\usage{
\S4method{marginal_pdf_2}{ContRV,ContRV,`function`}(crv1, crv2, dens)
}
\arguments{
\item{crv1}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{crv2}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{dens}{Funcția densitate comună a celor două variabile aleatoare}
}
\value{
Returnează o funcție, ce reprezintă densitatea marginală a lui y.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează densitatea marginală al lui y,
din două variabile aleatoare continue, utilizând funcția densitate comună acelor două.
}
\details{
Utilizează integrarea dupa o singură variabilă a celor două variabile pentru calcularea acestei densități.
}
\examples{
obj1 <- ContRV(pdf = function(x) { 1 }, lowerBound = 0, upperBound = 1)
obj2 <- ContRV(pdf = function(x) { 1 }, lowerBound = 1, upperBound = 2)
f2 <- marginal_pdf_2(obj1, obj2, function(x, y) 8/3 * x^3 *y)
f2(3) # 2
}
