% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{MAX_NUMERIC_LIMIT}
\alias{MAX_NUMERIC_LIMIT}
\title{Limita numerică maximă}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_NUMERIC_LIMIT
}
\description{
Setează limită numerică maximă constantă
}
\details{
Această constantă este specifică acestui proiect
}
\keyword{datasets}
