% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{mean,ContRV-method}
\alias{mean,ContRV-method}
\title{Media unei variabile aleatoare continue}
\usage{
\S4method{mean}{ContRV}(x)
}
\arguments{
\item{x}{Obiect de tip variabilă aleatoare continuă}
}
\value{
Returnează o valoare numerică, ce reprezintă media variabilei aleatoare continue.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează media variabilei aleatoare continue
primită ca parametru.
}
\details{
Utilizează integrarea pentru a calcula valoarea mediei variabilei aleatoare.
}
\examples{
var1 <- ContRV(pdf = function(x) { 1 / 3 }, lowerBound = 3, upperBound = 6)
mean(var1)  # 4.5

norm_5_4 = function(x) { 1 / (2 * sqrt(2 * pi)) * exp(-(x - 5)^2 / (2 * 4)) }
crv <- ContRV(pdf = norm_5_4, lowerBound = -Inf, upperBound = Inf)
mean(crv) # 5
}
