% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{mean_crv_func}
\alias{mean_crv_func}
\title{Funcție generică mean_crv_func}
\usage{
mean_crv_func(crv, g)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă}

\item{g}{Funcția continuă care primește ca argument variabila aleatoare continuă}
}
\value{
Returnează o valoare numerică, ce reprezintă media variabilei aleatoare continue obținute.
}
\description{
Funcție generică disponibilă pentru overloading
}
