% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{MIN_NUMERIC_LIMIT}
\alias{MIN_NUMERIC_LIMIT}
\title{Limita numerică minimă}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_NUMERIC_LIMIT
}
\description{
Setează limită numerică minimă constantă
}
\details{
Această constantă este specifică acestui proiect
}
\keyword{datasets}
