% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{moment_centrat,ContRV,numeric-method}
\alias{moment_centrat,ContRV,numeric-method}
\title{Momentul centrat al unei variabile aleatoare continue}
\usage{
\S4method{moment_centrat}{ContRV,numeric}(crv, order)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă}

\item{order}{Numarul de ordine al momentului centrat}
}
\value{
Returnează o valoare care reprezeintă momentul centrat de ordin order.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează momentul centrat
de ordin m al unei variabile aleatoare continue
}
\details{
Utilizează integrarea pentru a calcula valoarea momentului centrat al variabilei aleatoare.
}
\examples{
crv <- normal(5, 4)
moment_centrat(crv, 2) # 4 - Varianta
moment_centrat(crv, 4) # 48
}
