% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{moment_centrat}
\alias{moment_centrat}
\title{Funcția generică moment_centrat}
\usage{
moment_centrat(crv, order)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{order}{Numarul de ordine al momentului centrat}
}
\value{
Returnează o valoare care reprezeintă momentul centrat de ordin order.
}
\description{
Funcție generică disponibilă pentru overloading
}
