% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{moment_initial,ContRV,numeric-method}
\alias{moment_initial,ContRV,numeric-method}
\title{Momentul initial al unei variabile aleatoare continue}
\usage{
\S4method{moment_initial}{ContRV,numeric}(crv, order)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă}

\item{order}{Numarul de ordine al momentului centrat}
}
\value{
Returnează o valoare care reprezeintă momentul initial de ordin order.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează momentul initial
de ordin m al unei variabile aleatoare continue
}
\details{
Utilizează integrarea pentru a calcula valoarea momentului initial al variabilei aleatoare.
}
\examples{
crv <- normal(5, 4)
moment_initial(crv, 2) # 29
moment_initial(crv, 4) # 1273
}
