% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{normal}
\alias{normal}
\title{Repartitia normală}
\usage{
normal(mean, disp)
}
\arguments{
\item{mean}{Media repartiției normale}

\item{disp}{Dispersia repartiției normale}
}
\value{
Returnează un obiect de tip \code{ContRV} sau o eroare.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție normală
}
\details{
Repartiția normală are următoarele caracteristici:
\itemize{
    \item Parametru: \code{mean, disp > 0}
    \item Domeniul de valori: \code{R}
    \item Densitate: f(x) = 1 / (sd * sqrt(2*pi)) * e^(-(x - mean)^2/(2*sd^2))
    \item Repartiția: F(x) = 1/2 * [1 + erf((x - mean) / (sd * sqrt(2))]
    \item Media: mean (din parametrul funcției)
    \item Mediana: mean (din parametrul funcției)
    \item Dispersia: disp (din parametrul funcției)
}
Funcția returnează o eroare în cazul în care \code{disp <= 0} (dispersia nu poate fi nepozitivă). \cr
Repartițiile normale se utilizează în diferite domenii, în special în fizică (de exemplu:
funcția de densitate de probabilitate într-un oscilator armonic liniar (cuantic), poziția unei
particule care este în proces de difuziune).
}
\examples{
var1 <- normal(5, 2)
var2 <- normal(0, 1)
mean(var1) # 5
Var(var2) # 1
}
\references{
Normal distribution: https://en.wikipedia.org/wiki/Normal_distribution
}
