% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normalize_function}
\alias{normalize_function}
\title{Normalizarea unei funcții}
\usage{
normalize_function(func, lowerBound, upperBound)
}
\arguments{
\item{func}{Funcția pentru care trebuie să aflam funcția normalizată}

\item{lowerBound}{Limita inferioară de integrare a funcției}

\item{upperBound}{Limita superioară de integrare a funcției}
}
\value{
Returnează o funcție care reprezintă funcția normalizată sau \code{NULL} în caz de eroare
}
\description{
Fiind dată o funcție și limita superioara și inferioară,
calculează funcția normalizată (egala cu produsul dintre funcția inițială și
constanta de normalizare, dacă există) din această funcție pe intervalul indicat.
}
\details{
Erori la calcularea funcției normalizate pot apărea în două cazuri, când nu putem calcula constanta de normalizare:
\itemize{
    \item Valoarea integralei este 0 pe acel interval
    \item Integrala este divergentă pe acel interval
}
}
\examples{
func <- normalize_function(function(x) { 1 / x^2 }, 1, 2)
fx <- normalize_function(function(x) { x }, 0, 1)
}
