% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normalizing_constant}
\alias{normalizing_constant}
\title{Constanta de normalizare a integralei}
\usage{
normalizing_constant(func, lowerBound, upperBound)
}
\arguments{
\item{func}{Funcția pentru care trebuie să aflam constanta de normalizare}

\item{lowerBound}{Limita inferioară de integrare}

\item{upperBound}{Limita superioară de integrare}
}
\value{
Returnează un număr care reprezintă constanta de normalizare a funcției sau \code{NULL} în caz de eroare
}
\description{
Fiind dată o funcție și limita superioara și inferioară, află constanta de normalizare
(egala cu inversa integralei, daca exista si nu este nulă) din această funcție pe intervalul indicat.
}
\details{
Erori la calcularea constantei pot apărea în două cazuri:
\itemize{
    \item Valoarea integralei este 0 pe acel interval
    \item Integrala este divergentă pe acel interval
}
}
\examples{
const <- normalizing_constant(function(x) { 1/2 * x }, -2, 10)
K_norm <- normalizing_constant(function(x) { x }, 0, 2)
}
