% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Generarea graficelor funcției densitate de probabilitate și funcției de repartiție a unei repartiții
cunoscute}
\usage{
plot_distribution(name, x_int, y_int, ...)
}
\arguments{
\item{name}{Numele repartiției solicitate pentru afișarea graficului}

\item{x_int}{Intervalul de valori pentru argumentul repartiției}

\item{y_int}{Intervalul de valori pentru valorile functiei densitate și repartie}

\item{...}{Parametri pentru repartiția solicitată}
}
\value{
Returnează graficele funcțiilor densitate de probabilitate și repartiție a unei repartiții cunoscute.
}
\description{
Funcție care genereză graficele funcției densitate de probabilitate și
funcției de repartiție a unei repartiții cunoscute, trimise ca parametru.
}
\examples{
plot_distribution("uniform", x_int = c(0, 3), y_int = c(0, 2), a = 1, b = 2)
plot_distribution("exponential", x_int = c(0, 3), y_int = c(0, 2), lambda = 2)
plot_distribution("normal", x_int = c(-1, 10), y_int = c(0, 1), mean = 5, disp = 2)
plot_distribution("f", x_int = c(0, 5), y_int = c(0, 1), 5, 2)
}
