% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{plot_RV}
\alias{plot_RV}
\title{Funcția generică plot.RV}
\usage{
plot_RV(crv, x_int, y_int, color)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{x_int}{Intervalul de valori pentru argumentul functiei}

\item{y_int}{Intervalul de valori pentru valorile functiei densitate și repartie}

\item{color}{Culoarea graficului}
}
\value{
Returnează graficul funcției densitate de probabilitate și funcției de repartiție.
}
\description{
Funcție generică disponibilă pentru overloading
}
