% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{repf}
\alias{repf}
\title{Repartitia Fisher (F)}
\usage{
repf(d1, d2)
}
\arguments{
\item{d1}{Gradul de liberate 1}

\item{d2}{Gradul de libarate 2}
}
\value{
Returnează un obiect de tip \code{ContRV} sau o eroare.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție F
}
\details{
Repartiția normală are următoarele caracteristici:
\itemize{
    \item Parametru: \code{mean, disp > 0}
    \item Domeniul de valori: \code{R}
    \item Densitate: f(x) = sqrt(((d1*x)^d1 * d2^d2) / ((d1*x + d2)^(d1 + d2))) / (x * beta(d1 / 2, d2 / 2))
    \item Media: d2 / (d2 - 2)
    \item Dispersia: 2 * d2^2 * (d1 + d2 - 2) / (d1 * (d2 - 2)^2 * (d2 - 4))
}
Funcția returnează o eroare în cazul în care \code{d1 <= 0} și \code{d2 <= 0}
(gradele de liberate nu pot fi nenegative). \cr
Un exemplu de utilizare a repartiției Fisher este distrubuția veniturilor. Pentru mai multe detalii,
vizitați linkurile din referința.
}
\examples{
var1 <- repf(5, 2)
var2 <- repf(1, 1)
}
\references{
F-distribution: https://en.wikipedia.org/wiki/F-distribution \cr
When to use F-distribution: https://www.dummies.com/education/math/business-statistics/when-to-use-the-f-distribution/
}
