% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{repgamma}
\alias{repgamma}
\title{Repartiția Gamma}
\usage{
repgamma(eta, lambda)
}
\arguments{
\item{eta}{parametru de formă}

\item{lambda}{repartiție exponențială}
}
\value{
Returnează un obiect de tip \code{ContRV}.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție Gamma.
}
\details{
Repartiția Gamma are următoarele caracteristici:
\itemize{
    \item Parametri: \code{eta > 0, lambda > 0}
    \item Domeniul de valori: \code{(0, Inf)}
    \item Densitate: \eqn{ f(x) = (lambda^eta)/gamma(eta) * (x^(eta - 1)) * exp(-lambda * x)}
    \item Aplicatie: Repartiția Gamma reprezintă repartiția probabilității entropiei maxime.
}
Funcția returnează o eroare în cazul în care \eqn{eta <= 0} sau \eqn{lambda <= 0}
}
\examples{
var1 <- repgamma(9, 2)
# var2 <- repgamma(-100, 2) # Eroare: -100 <= 0
# var3 <- repgamma(9, -4) # Eroare: -4 <= 0
}
\references{
Gamma distribution: https://en.wikipedia.org/wiki/Gamma_distribution
}
