% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{student}
\alias{student}
\title{Repartiția Student}
\usage{
student(niu)
}
\arguments{
\item{niu}{gradul de libertate}
}
\value{
Returnează un obiect de tip \code{ContRV}.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție Student.
}
\details{
Repartiția Student are următoarele caracteristici:
\itemize{
    \item Parametri: \code{niu > 0}
    \item Domeniul de valori: \code{R}
    \item Densitate: \eqn{ f(x) = (gamma((niu + 1) / 2)) / (sqrt(pi * niu) * gamma(niu / 2)) * ((1 + (x^2) / niu)^(-(niu + 1)/2))}
    \item Aplicatie: Repartiția Student, la fel ca și distribuția normală, are formă de clopot, insă are capetele mai pronunțate, deci e mult mai folositoare pentru modelarea distribuțiilor a căror valori pot cădea mai departe de medie.
}
Funcția returnează o eroare în cazul în care \eqn{niu <= 0}
}
\examples{
var1 <- student(1)
# var2 <- student(-1) # Eroare: -1 <= 0
}
\references{
Student distribution: https://en.wikipedia.org/wiki/Student%27s_t-distribution
}
