% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{uniform}
\alias{uniform}
\title{Repartitia uniforma}
\usage{
uniform(a, b)
}
\arguments{
\item{a}{Limita inferioară}

\item{b}{Limita superioară}
}
\value{
Returnează un obiect de tip \code{ContRV} sau o eroare.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție uniformă.
}
\details{
Repartiția uniformă are următoarele caracteristici:
\itemize{
    \item Parametri: \code{a < b}
    \item Domeniul de valori: \code{[a, b]}
    \item Densitate: f(x) = 1 / (b - a)
    \item Repartiția: F(x) = (x - a) / (b - a)
    \item Media: 1/2 * (a + b)
    \item Mediana: 1/2 * (a + b)
    \item Dispersia: 1/12 * (b - a)^2
}
Funcția returnează o eroare în cazul în care \code{a >= b}. \cr
Toate rezultatele din domeniul de valori au probabilitate egală. Repartiția uniformă
este utilizată destul de des, din motiv că probabilități pentru aceasta sunt ușor de calculat.
În practică, acestea sunt utilizate destul de des în următoarele aplicații: testarea
ipotezelor, finanțe, studii în fizică (de exemplu: emisiile de particule radioactive),
în economie (cerere și ofertă).
}
\examples{
var1 <- uniform(2, 5)
# var2 <- uniform(5, 2) # Eroare: 5 > 2
}
\references{
Continuous uniform distribution: https://en.wikipedia.org/wiki/Continuous_uniform_distribution \cr
Uniform Distributions: https://online.stat.psu.edu/stat414/lesson/14/14.6
}
