% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{Var,ContRV-method}
\alias{Var,ContRV-method}
\title{Dispersia unei variabile aleatoare continue}
\usage{
\S4method{Var}{ContRV}(crv)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă}
}
\value{
Returnează o valoare numerică, ce reprezintă varianța variabilei aleatoare continue.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează dispersia variabilei aleatoare continue
primită ca parametru.
}
\details{
Utilizează integrarea pentru a calcula valoarea varianței variabilei aleatoare.
}
\examples{
var1 <- ContRV(pdf = function(x) { 1 / 3 }, lowerBound = 3, upperBound = 6)
Var(var1)  # 0.75

norm_5_4 = function(x) { 1 / (2 * sqrt(2 * pi)) * exp(-(x - 5)^2 / (2 * 4)) }
crv <- ContRV(pdf = norm_5_4, lowerBound = -Inf, upperBound = Inf)
Var(crv) # 4
}
