% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{Var}
\alias{Var}
\title{Funcția generică Var}
\usage{
Var(crv)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}
}
\value{
Returnează o valoare care reprezintă varianța obiectului trimis ca parametru.
}
\description{
Funcție generică disponibilă pentru overloading
}
