% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{var_crv_func,ContRV,function-method}
\alias{var_crv_func,ContRV,function-method}
\title{Dispersia unei variabile aleatoare continue, asupra căreia este aplicată o funcție}
\usage{
\S4method{var_crv_func}{ContRV,`function`}(crv, g)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă}

\item{g}{Funcția continuă care primește ca argument variabila aleatoare continuă}
}
\value{
Returnează o valoare numerică, ce reprezintă dispersia variabilei aleatoare continue obținute.
}
\description{
Metodă a clasei \code{ContRV} care calculează și returnează varianța variabilei aleatoare continue,
asupra căreia este aplicată o funție continuă primită ca parametru.
}
\details{
Utilizează formula varianței pentru a calcula valoarea dispersiei variabilei aleatoare obținute.
}
\examples{
var1 <- ContRV(pdf = function(x) { 2 * exp(-2 * x) }, lowerBound = 0, upperBound = Inf)
var_crv_func(var1, function(x){ x ^ 2 })
}
