% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContRV.R
\name{var_crv_func}
\alias{var_crv_func}
\title{Funcția generică var_crv_func}
\usage{
var_crv_func(crv, g)
}
\arguments{
\item{crv}{Obiect de tip variabilă aleatoare continuă sau altul, în dependență de implementare}

\item{g}{Funcția continuă care primește ca argument variabila aleatoare continuă}
}
\value{
Returnează o valoare care reprezintă varianța obiectului trimis ca parametru.
}
\description{
Funcție generică disponibilă pentru overloading
}
