% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepartiiClasice.R
\name{weibull}
\alias{weibull}
\title{Repartiția Weibull}
\usage{
weibull(beta, gama, eta)
}
\arguments{
\item{beta}{parametru de formă}

\item{gama}{parametru de scară}

\item{eta}{parametru de poziție sau al originii de timp}
}
\value{
Returnează un obiect de tip \code{ContRV}.
}
\description{
Funcție care generează o variabilă aleatoare continuă ce are repartiție Weibull.
}
\details{
Repartiția Weibull are următoarele caracteristici:
\itemize{
    \item Parametri: \code{beta > 0, gama, eta > 0}
    \item Domeniul de valori: \code{[gama, Inf)}
    \item Densitate: \eqn{ f(x) = beta/eta * ((x - gama)/eta)^(beta - 1) * exp(-((x - gama)/eta)^beta)}
    \item Aplicatie: Repartiția Weibull este des utilizată în analiza fiabilității unui product.
}
Funcția returnează o eroare în cazul în care \eqn{beta <= 0} sau \eqn{eta <= 0}
}
\examples{
var1 <- weibull(3,0,200)
# var2 <- weibull(-3,40, 3) # Eroare: -3 <= 0
# var3 <- weibull(3,40, -4) # Eroare: -4 <= 0
}
\references{
Weibull distribution: https://en.wikipedia.org/wiki/Weibull_distribution
}
